Refactored appointment API calls into two files.
